Refactor file handling and storage in Files Service
The namespace of Files Service was updated in several files. The appsettings.json and program.cs files were updated to handle file resources using a new root path system. A new file handler was added and has been included in the IoC container. In addition, a new AutoMapper profile was introduced to map entities and their corresponding DTO's.